+++ codepen-clone/.eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx":"1","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx":"2","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx":"3","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/unpkg-path-plugin.ts":"4"},{"size":249,"mtime":1623241596645,"results":"5","hashOfConfig":"6"},{"size":1442,"mtime":1624466487030,"results":"7","hashOfConfig":"6"},{"size":1742,"mtime":1623237859200,"results":"8","hashOfConfig":"6"},{"size":2067,"mtime":1624468263458,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ssvdrz",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/unpkg-path-plugin.ts",[]]
+[{"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx":"1","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx":"2","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx":"3","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/unpkg-path-plugin.ts":"4","/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/fetch-plugin.ts":"5"},{"size":249,"mtime":1623241596645,"results":"6","hashOfConfig":"7"},{"size":1511,"mtime":1624469089695,"results":"8","hashOfConfig":"7"},{"size":1742,"mtime":1623237859200,"results":"9","hashOfConfig":"7"},{"size":826,"mtime":1624468967693,"results":"10","hashOfConfig":"7"},{"size":1220,"mtime":1624468882052,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ssvdrz",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/index.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/App.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/components/code-editor.tsx",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/unpkg-path-plugin.ts",[],"/Users/gabriel/Documents/code/codepen-clone-typescript-react/codepen-clone-typescript/codepen-clone/src/plugins/fetch-plugin.ts",[]]
