+++ codepen-clone/src/plugins/unpkg-path-plugin.ts
@@ -1,21 +1,6 @@
 import * as esbuild from "esbuild-wasm";
-import axios from "axios";
-import localForage from "localforage";
 
-const fileCache = localForage.createInstance({
-  name: "filecache",
-});
-//
-
-(async () => {
-  // save sth to indexDB
-  await fileCache.setItem("color", "blue");
-  // read from indexDB
-  const color = await fileCache.getItem("color");
-  console.log(color);
-})();
-
-export const unpkgPathPlugin = (inputCode: string) => {
+export const unpkgPathPlugin = () => {
   return {
     name: "unpkg-path-plugin",
     setup(build: esbuild.PluginBuild) {
@@ -39,39 +24,6 @@ export const unpkgPathPlugin = (inputCode: string) => {
           namespace: "a",
           path: `https://unpkg.com/${args.path}`,
         };
-      });
-
-      build.onLoad({ filter: /.*/ }, async (args: any) => {
-        console.log("onLoad", args);
-
-        if (args.path === "index.js") {
-          return {
-            loader: "jsx",
-            contents: inputCode,
-          };
-        }
-
-        // check if already fetched this file
-        const cachedResult = await fileCache.getItem<esbuild.OnLoadResult>(
-          args.path
-        );
-        if (cachedResult) {
-          return cachedResult;
-        }
-
-        // if its in cache already return
-        // else request file
-        const { data, request } = await axios.get(args.path);
-
-        const result: esbuild.OnLoadResult = {
-          loader: "jsx",
-          contents: data,
-          resolveDir: new URL("./", request.responseURL).pathname,
-        };
-
-        // store response in cache
-        await fileCache.setItem(args.path, result);
-        return result;
       });
     },
   };
